Support img extraction.
The UrlExtractor can work for img tags as well. All we need to do is allow the attribute name to be specified. Additionally, building regex is expensive. We can eliminate how often it is created by making the base address part of the getUrls method, rather than passing it to the ctor. Then we can declare the extractors outside the loop.