Add -Wall compiler option.
I was hoping adding some compiler warnings would reveal the cause of my ominous core dump that only showed up after running for 12+ hours. For a long time, I was convinced there was a bug with the way I was handling exceptions. Adding the -Wall option helped me discover that I was not returning a value from PageDownloader::download when an exception was thrown. So, I was invariable returning garbage that eventually caused a core dump.