Create thread pool class to limit number of threads.
The plan is to make the spider multithreaded. This will allow the spider to download many pages and files all at once. Since a single task can result in many new tasks, we need a way to managing how many threads are run at one time. We also need a convenient way to detect when the spider process completes. The best approach is to keep track of a counter, incrementing it for each task that needs to run and decrementing when they complete. When the last task completes, the counter goes to zero, which will unlock a mutex, allowing the main thread to run to completion. The thread pool class is the first piece needed to start going down this path. This commit provides the general idea for a task pool, but I have not thoroughly tested whether its implementation is correct or optimal.