Lock task queue initially.
I was not locking the task queue initially. On a virtual machine, the sofware was too slow to start up a thread and the queue was filled before the first consumer waiting on the `has_tasks_mutex`. Running off the physical hardware, the race condition was discovered. I changed the `has_tasks_mutex` to be locked to start with.

I now determine the number of threads based on the number of processors/cores/hyperthreads/etc.

I do not download files that already exist in the file system.